Add project`s exceptions. Delete PrivateBankException.java. Change path for pbapi, rename class to following project structure.
